+++ src/components/Cards/Power/PowerCard.tsx
@@ -55,11 +55,13 @@ export default ({
 
     const [internalVTerm, setInternalVTerm] = useState(charger.vTerm);
     const [internalIChg, setInternalIChg] = useState(charger.iChg);
+    const [internalBatLim, setInternalBatLim] = useState(charger.batLim);
 
     // NumberInputSliderWithUnit do not use charger.<prop> as value as we send only at on change complete
     useEffect(() => {
         setInternalVTerm(charger.vTerm);
         setInternalIChg(charger.iChg);
+        setInternalBatLim(charger.batLim);
     }, [charger]);
 
     return (
@@ -129,6 +131,20 @@ export default ({
 
             {!summary && (
                 <>
+                    <NumberInputSliderWithUnit
+                        label={
+                            <div>
+                                <span>BAT</span>
+                                <span className="subscript">LIM</span>
+                            </div>
+                        }
+                        unit="mA"
+                        disabled={disabled}
+                        range={npmDevice.getChargerIBatLimRange()}
+                        value={internalBatLim}
+                        onChange={setInternalBatLim}
+                        onChangeComplete={v => npmDevice.setChargerBatLim(v)}
+                    />
                     <Toggle
                         label={
                             <DocumentationTooltip