+++ src/features/pmicControl/npm/npm1300/tests/offlineSetters.test.ts
@@ -60,6 +60,13 @@ describe('PMIC 1300 - Setters Offline tests', () => {
         expect(mockOnChargerUpdate).toBeCalledWith({ iTerm: '10%' });
     });
 
+    test('Set setChargerBatLim', async () => {
+        await pmic.setChargerBatLim(1000);
+
+        expect(mockOnChargerUpdate).toBeCalledTimes(1);
+        expect(mockOnChargerUpdate).toBeCalledWith({ batLim: 1000 });
+    });
+
     test('Set setChargerEnabledRecharging ', async () => {
         await pmic.setChargerEnabledRecharging(true);